requirements.txt
# Django

asgiref==3.2.10
Django==3.0.8
pytz==2020.1
sqlparse==0.3.1

# Decouple
python-decouple==3.3

# dj-database-url
dj-database-url==0.5.0

# Pillow
Pillow==7.2.0

# Heroku
django-heroku==0.3.1
gunicorn==20.0.4
psycopg2-binary==2.8.5
psycopg2==2.8.5
whitenoise==5.1.0

# Cloudinary
certifi==2020.6.20
chardet==3.0.4
cloudinary==1.21.1
dj3-cloudinary-storage==0.0.3
idna==2.10
requests==2.24.0
six==1.15.0
urllib3==1.25.9